feat: macOS support
