+++ darwin/push.podspec
@@ -14,8 +14,11 @@ A new flutter plugin project.
   s.author           = { 'Ben Butterworth' => 'push@orth.uk' }
   s.source           = { :path => '.' }
   s.source_files = 'Classes/**/*'
-  s.dependency 'Flutter'
-  s.platform = :ios, '11.0'
+  # As per https://docs.flutter.dev/packages-and-plugins/developing-packages#shared-ios-and-macos-implementations
+  s.ios.dependency 'Flutter'
+  s.osx.dependency 'FlutterMacOS'
+  s.ios.deployment_target = '11.0'
+  s.osx.deployment_target = '10.14'
 
   # Flutter.framework does not contain a i386 slice.
   s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES', 'EXCLUDED_ARCHS[sdk=iphonesimulator*]' => 'i386' }
