--- ios/Classes/PushPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface PushPlugin : NSObject<FlutterPlugin>
-@end
