+++ mise.toml
@@ -1,6 +1,5 @@
 [tasks.lint]
 description = "Lint"
-dir = "push"
 run = [
   'dart format --set-exit-if-changed .',
   'swiftformat --lint --exclude example/macos/Flutter/GeneratedPluginRegistrant.swift .',
@@ -8,7 +7,6 @@ run = [
 
 [tasks.format]
 description = "Format"
-dir = "push"
 run = [
   # https://dart.dev/tools/dart-format
   'dart format .', # https://github.com/nicklockwood/SwiftFormat, not https://github.com/apple/swift-format
@@ -17,7 +15,6 @@ run = [
 
 [tasks.codegen]
 description = "Generate code"
-dir = "push"
 run = [
   'dart run pigeon --input pigeons/push_api.dart',
 ]
