+++ tools/src/index.ts
@@ -1,28 +1,29 @@
+import { z } from 'zod';
 import { sendAndroid } from './android';
 import { config } from './config';
-import { sendIos } from './ios';
+import { sendDarwin } from './darwin';
 import { program } from 'commander';
 
 program.name('push')
     .description('Send push notification to iOS and Android devices by their token.')
     .option('-b, --background', 'Send iOS background message', false)
     .option('-i, --ios', 'Send iOS push notification', false)
+    .option('-m, --macos', 'Send macOS push notification', false)
     .option('-a, --android', 'Send Android push notification', false)
     .action(async (options) => {
-        if (!options.ios && !options.android) {
-            console.error('You must specify at least one of --ios or --android');
-            process.exit(1);
+        const runAllPlatforms = !options.ios && !options.macos && !options.android;
+        if (runAllPlatforms) {
+            console.log("Sending to all devices since no platform was specified");
         }
-        if (options.ios) {
-            if (options.background) {
-                await sendIos(config.apple_apns, "background");
-            } else {
-                await sendIos(config.apple_apns, "alert");
-            }
+        if (runAllPlatforms || options.ios) {
+            const deviceToken = z.string().parse(config.apple_apns?.ios_device_token);
+            await sendDarwin(config.apple_apns, options.background ? "background" : "alert", deviceToken);
         }
-        if (options.android) {
+        if (runAllPlatforms || options.macos) {
+            const deviceToken = z.string().parse(config.apple_apns?.macos_device_token);
+            await sendDarwin(config.apple_apns, options.background ? "background" : "alert", deviceToken);
+        }
+        if (runAllPlatforms || options.android) {
             await sendAndroid(config.google_fcm);
         }
-    });
-
-program.parse();
+    }).parse();
